+++ package.json
@@ -18,8 +18,7 @@
     "react-select": "^1.0.0-rc.5",
     "redux": "^3.7.0",
     "redux-saga": "^0.15.4",
-    "regex": "^0.1.1",
-    "schema-inspector": "^1.6.8"
+    "regex": "^0.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
