--- src/components/submission/ArmorSetForm.js
@@ -1,5 +1,0 @@
-import React from "react"
-
-class ArmorSetForm extends React.Component {
-
-}
