--- src/components/submission/SubmissionForm.css
@@ -1,9 +1,0 @@
-.submission-form--container {
-  display: flex;
-  flex-direction: row;
-}
-
-.submission-form--body {
-  display: flex;
-  flex-direction: row;
-}
