--- src/components/ui/SearchSelectionInput.css
@@ -1,18 +1,0 @@
-.search-menu {
-  max-height: 100px;
-  overflow-y: scroll;
-}
-
-.search-menu--input {
-  width: 100%;
-}
-
-.search-menu--item {
-  background-color: white;
-  border-block-start: 0 solid white;
-  border-top: 1px solid black;
-}
-
-.hidden {
-  visibility: hidden;
-}
